feat: 拆分适配浏览器内置webmcp功能，优化代码结构
